Add another model and a ForeignKey to it for testing.
Fix the tests broken by the addition of an extra field.

Fix a bug in the mixin where it would get the list of header fields from one
source, and the list of fields to extract data from a different source (one
was queryset.values_list().fields, the other was model._meta.get_all_field_names)
causing the headings and values not to line up.

Add support for using a values_list queryset with a defined list of fields,
and tests for it.