Pass the model, rather than the data, to generate_headers().
It used to just extract the model from the data, but this makes testing harder
because you can't add a 'model' attribute to a plain list(). And it
unnecessarily hides the fact that we actually need a model to calculate the
headers, and we don't care about what's in the rest of the data.

Also make the 'fields' parameter of generate_headers required, not optional.
If you don't know what fields to use, call get_fields(). I don't think anyone
is actually using this, because nobody should really be calling
generate_headers() directly.