Add support for calculated fields.
Pass a callable name (a view method or global function) as a field, and
give it a 'fields' attribute which is a list of fields that it requires from
the database. These fields will be passed to values_list(), and the method
will be called with the values, once per row, to return a calculated value
based on them, which goes into the exported (Excel/CSV) output.